moved outline and courseevent to reflect dependencies
