+++ src/main/java/entities/course/outline/OutlineFactory.java
@@ -1,4 +1,4 @@
-package entities.outline;
+package entities.course.outline;
 
 public class OutlineFactory implements OutlineFactoryInterface{
     public Outline create() {
