+++ src/main/java/entities/course/outline/OutlineFactoryInterface.java
@@ -1,4 +1,4 @@
-package entities.outline;
+package entities.course.outline;
 
 public interface OutlineFactoryInterface {
     Outline create();
