+++ src/test/java/courseManager/CourseManagerTest.java
@@ -3,7 +3,7 @@ package courseManager;
 import entities.assessment.Assessment;
 import entities.assessment.assessmentInstance.AssessmentInstance;
 import entities.course.Course;
-import entities.courseEvent.CourseEvent;
+import entities.course.courseEvent.CourseEvent;
 import entities.assessment.instanceList.InstanceList;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
